refactor: extract SystemDiagnosticsInfo.swift from DebugBundleExporter.swift
Extracts the `SystemDiagnosticsInfo` enum (currentArchitecture uname
wrapper) from DebugBundleExporter.swift into its own file.

Byte-identical extraction of lines 98-115. Adds `import Darwin` (for
utsname/uname) and `import Foundation` (for whitespacesAndNewlines).

Reduces DebugBundleExporter.swift from 115 to 96 lines. Tests: 667.
